requirements.txt
Django==4.2.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-jazzmin==3.0.1
django-ckeditor==6.6.1
Pillow==10.0.0
psycopg2-binary==2.9.6
python-decouple==3.8
